Moved main function to separate class.
