+++ BNA.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.421
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BNAC", "BNAC\BNAC.csproj", "{4B74E9C7-1FE3-4306-991C-76C9C6F146E7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BNA", "BNAC\BNA.csproj", "{4B74E9C7-1FE3-4306-991C-76C9C6F146E7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
