+++ BNAC/Token.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace BNAC
+namespace BNA
 {
 	internal enum Keyword
 	{
